requirements.txt
streamlit
crewai
openai
requests
python-dotenv
